Correctly use stimulus to attach my javascript instead.
