+++ app/views/trucks/show.html.erb
@@ -13,5 +13,4 @@
       <%= button_to "Destroy this truck", @truck, method: :delete %>
     <% end %>
   </div>
-<script type="module" src="/assets/controllers/notifications.js"></script>
-
+<span data-controller="trucks--show"></span>
